Fix message memleak, don't mutate global maxListeners, various other minor fixes.
Changing `EventEmitter.defaultMaxListeners` changes it for every EventEmitter in the program.
Fixed it so that it only changes the listeners for the socket.
Set it to automatically remove the message handler after first `'message'` event by using `.once` instead of `.on`.
Set it up so it removes the `message` handler after the `stunTest` promise resolves for any reason.
Set it up so it clears the `bgOp` timeout after the `stunTest` promise resolves for any reason.
Might be that setting `maxListeners` or collecting `backgroundOps` is not even required with the above changes.
Stopped it from retrying `sendMessage` after 3 tries. Looked like there was a missing `return` statement after `resolve`.
Stopped `sendMessage` from being a global. Looked like a missing `var`/`let`/`const`.
Defined `ERROR`.